image hierarchy , output on CLI
image hierarchy and output on the command line . The data from the folder "train" given in kaggle, is split into training and testing. Confusion matrix is plotted both for training and testing data.   

later tested the code of new set of images in folder "test_stg1" . I have not plotted confusion matrix for these as we do not have labels for new set of images.